compareTo; weight is integer

+++ Esercizi/Esercizio1/Ordinamento/src/esercizio4/Graph.java
@@ -81,8 +81,8 @@ public abstract class Graph {
     /**
      * Restituisce la somma dei pesi dei collegamenti tra i nodi del grafo.
      */
-    public double weight(){
-        double total = 0;
+    public Integer weight(){
+        Integer total = 0;
         for(Node n : this.nodes)
             total += n.weight();
         return total;
